pyproject.toml
[project]
name = "github-webhooks"
version = "0.1.0"
description = "TechStaX Webhook Receiver"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "flask-pymongo>=3.0.1",
    "gunicorn>=24.1.1",
    "pymongo>=4.16.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
]
